disentangle the server from the handler
