+++ pegasus-0.0.1-5.rockspec
@@ -24,6 +24,7 @@ build = {
   type = "builtin",
   modules = {
     ['pegasus.init']     = "src/pegasus/init.lua",
+    ['pegasus.handler']  = 'src/pegasus/handler.lua',
     ['pegasus.request']  = 'src/pegasus/request.lua',
     ['pegasus.response'] = 'src/pegasus/response.lua',
     ['pegasus.file']     = 'src/pegasus/file.lua'
